fixed DepartmentControllers

+++ src/modules/department/department.controllers.ts
@@ -7,12 +7,16 @@ import {AppErrorUtil} from "../../utils/AppError.util";
 const departmentService = new DepartmentServices();
 
 export class DepartmentControllers {
+    getAll = catchAsyncUtil(async (_req, res: Response) => {
+        const departments = await departmentService.getAll();
+        res.api.success(200, "OK", departments);
+    })
+
     getById = catchAsyncUtil(async (req: Request, res: Response) => {
         const id = parseInt(req.params.id);
         if (isNaN(id)) throw new AppErrorUtil(400, "Invalid department ID");
 
         const department = await departmentService.getById(id);
-
         if (!department) throw new AppErrorUtil(404, "Department not found");
 
         res.api.success(200, "OK", department);
@@ -21,11 +25,6 @@ export class DepartmentControllers {
     create = catchAsyncUtil(async (req: Request, res: Response) => {
         const department = await departmentService.create(req.body);
         res.api.success(201, "Created", department);
-    })
-
-    getAll = catchAsyncUtil(async (_req, res: Response) => {
-        const departments = await departmentService.getAll();
-        res.api.success(200, "OK", departments);
     })
 
     update = catchAsyncUtil(async (req: Request, res: Response) => {
@@ -40,7 +39,10 @@ export class DepartmentControllers {
 
     delete = catchAsyncUtil(async (req: Request, res: Response) => {
         const id = parseInt(req.params.id);
+        if (isNaN(id)) throw new AppErrorUtil(400, "Invalid department ID");
+
         await departmentService.delete(id);
+
         res.api.success(204, "Deleted");
     })
 }
